fix(calculator): caller import + stub returns failed CF
- SnapshotChunkProcessor: add kotlinx.coroutines.future.await import
  for CompletableFuture.await() bridge. Convert resultCount Long to
  Int for ChunkResult.
- StubObjectStorage: catch synchronous exceptions in handlePutStreamMultipart
  and return CompletableFuture.failedFuture. Matches the interface
  contract (putStreamMultipart must return a CF, not throw).

+++ module-calculator/src/main/kotlin/maple/calculator/processor/SnapshotChunkProcessor.kt
@@ -9,6 +9,7 @@ import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.future.await
 import kotlinx.coroutines.launch
 import maple.calculator.config.CoroutineDispatcherConverter
 import maple.calculator.config.PipelineProperties
@@ -102,7 +103,7 @@ class SnapshotChunkProcessor(
             calculatedCount = calculatedCount.get(),
             errorCount = errorCount.get(),
             resultObjectKey = writeResult.objectKey,
-            resultCount = writeResult.resultCount,
+            resultCount = writeResult.resultCount.toInt(),
             resultUncompressedBytes = writeResult.uncompressedBytes,
             resultCompressedBytes = writeResult.compressedBytes,
         )
